72-adding-team-member-functionalities

+++ src/main/java/com/example/rev_task_management_project02/models/User.java
@@ -34,7 +34,7 @@ public class User {
     private String phone;
 
     @Column(name = "manager_id")
-    private Integer managerId; // Optional relationship
+    private long managerId;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "status", nullable = false, length = 10)
